Splits current CalculateCoverageComponents integration test into one using Spark and another using Apache common for SVD.
This is in response to a difficult to reproduce error when testing in Travis. If the Spark
version is the one that has the problems it makes it more likely to be a race-condition problem.

The problem has been reported in issue #242

After a few Travis re-run the problem only occurs when using Spark SVD.

For now I disable the possibility of using Spark for the SVD/PCA.