HaplotypeCallerSpark (#616)
First HaplotypeCallerSpark implementation.  

This passes concordance tests but is not a perfect match with the non-spark version.  

This has some limitations vs the non-spark HaplotypeCaller.  Specifically, it doesn't yet support loading variant tracks as side inputs or writing output bams.  

Output is done on a single node so performance will be poor with large callsets.